requirements.txt
yfinance
pandas
numpy
kafka-python
psycopg2-binary
scripts